Add readable(path) condition function
It returns true if the path is a directory that can be read, or if
it is not a directory and can be opened as a file in read-only mode.

This is being added to allow conditions to guard against trying to
read the version of an executable that exists but is unreadable,
e.g. any of the official game executables for games installed
through the Microsoft Store or Xbox app.

The Windows test for evaling with an unreadable file doesn't use a
file that's unreadable for the same reason as what's motivating this
new function, but the error that is handled when checking the file
is the same kind (permission denied).