Avoid constructing paths using string literals
To avoid the chance of an issue involving a case-sensitive filesystem
and hardcoded string literals in loot-condition-interpreter.

However, I'm pretty sure this was a waste of time, because the library
input strings come from condition strings that do not necessarily
reflect the casing of filenames and paths in the filesystem that the
conditions are evaluated against, and that's much more likely to cause
issues.